refactor: move details screen to common module
