+++ app/lib/common/pages/medication_details.dart
