+++ app/lib/search/pages/cubit.dart
@@ -2,8 +2,8 @@ import 'dart:async';
 
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:http/http.dart' as http;
-
-import '../../../common/module.dart';
+import '../../../common/constants.dart';
+import '../../common/models/medication.dart';
 
 part 'cubit.freezed.dart';
